fix(ResliceCursorWidget): single App-level instance
Move instantiation of vtkResliceCursorWidget to App component
so that only one is created and shared between the oblique views.
